Configuration for YR
