+++ src/TSMapEditor/Config/Default/BaseMap.ini
@@ -8,7 +8,7 @@ NextScenario=
 AltNextScenario=
 Name=No name
 Author=Unknown Author
-GameMode=Custom Map
+GameMode=standard
 NewINIFormat=4
 CarryOverCap=0
 EndOfGame=no
@@ -27,23 +27,30 @@ MultiplayerOnly=0
 TiberiumGrowthEnabled=yes
 VeinGrowthEnabled=no
 IceGrowthEnabled=no
-TiberiumDeathToVisceroid=yes
+TiberiumDeathToVisceroid=no
 FreeRadar=no
 InitTime=10000
 
 [Lighting]
-Ambient=0.950
+Ambient=1.000
 Red=1.000
 Green=1.000
 Blue=1.000
 Ground=0.000
 Level=0.032
-IonAmbient=0.500
-IonRed=1.500
+IonAmbient=1.000
+IonRed=1.000
 IonGreen=1.000
-IonBlue=0.700
+IonBlue=1.000
 IonGround=0.000
 IonLevel=0.032
+DominatorAmbient=1.500
+DominatorAmbientChangeRate=0.009
+DominatorRed=0.850
+DominatorGreen=0.200
+DominatorBlue=0.300
+DominatorGround=0.000
+DominatorLevel=0.000
 
 [SpecialFlags]
 TiberiumGrows=yes
