+++ src/TSMapEditor/Config/Default/Bridges.ini
@@ -5,11 +5,21 @@
 
 
 [Bridges]
-0=Low Bridge
-;1=High Bridge
-;2=Railroad Bridge
+0=Low Concrete Bridge
+1=Low Wooden Bridge
+2=High Concrete Bridge
+3=High Wooden Bridge
 
-[Low Bridge]
+[Low Concrete Bridge]
+Kind=Low
+BridgeStart.EW=LOBRDB21
+BridgeEnd.EW=LOBRDB19
+BridgePieces.EW=LOBRDB01,LOBRDB02,LOBRDB03,LOBRDB04
+BridgeStart.NS=LOBRDB23
+BridgeEnd.NS=LOBRDB25
+BridgePieces.NS=LOBRDB10,LOBRDB11,LOBRDB12,LOBRDB13
+
+[Low Wooden Bridge]
 Kind=Low
 BridgeStart.EW=LOBRDG21
 BridgeEnd.EW=LOBRDG19
@@ -18,14 +28,14 @@ BridgeStart.NS=LOBRDG23
 BridgeEnd.NS=LOBRDG25
 BridgePieces.NS=LOBRDG10,LOBRDG11,LOBRDG12,LOBRDG13
 
-[High Bridge]
+[High Concrete Bridge]
 Kind=High
 BridgePieces.EW=BRIDGE1
 BridgePieces.NS=BRIDGE2
 TileSet=19
 
-[Railroad Bridge]
+[High Wooden Bridge]
 Kind=High
-BridgePieces.EW=RAILBRDG1
-BridgePieces.NS=RAILBRDG2
-TileSet=37
+BridgePieces.EW=BRIDGEB1
+BridgePieces.NS=BRIDGEB2
+TileSet=80
