+++ src/TSMapEditor/Config/Default/Constants.ini
@@ -1,4 +1,4 @@
-﻿; C&C World-Altering Editor (WAE)
+; C&C World-Altering Editor (WAE)
 ; https://github.com/CnCNet/WorldAlteringEditor
 
 ; This file defines various constants used in the editor.
@@ -6,17 +6,17 @@
 [Constants]
 
 ; Cell Size. 48x24 for TS, 60x30 for RA2.
-CellSizeX=48
-CellSizeY=24
+CellSizeX=60
+CellSizeY=30
 
 ; How much memory is reserved for the colors of a single sub-tile of a full tile.
-; A TS sub-tile contains 576 pixels, adjust it for RA2. Note that the editor can crash 
+; A TS sub-tile contains 576 pixels, RA2 one contains 900 pixels. Note that the editor can crash
 ; if a tile has fewer pixels than this.
-TileColorBufferSize=576
+TileColorBufferSize=900
 
 ; The maximum number of waypoints that the target game supports.
-; 100 in vanilla TS.
-MaxWaypoint=702
+; 100 in vanilla TS and RA2, 701 in YR.
+MaxWaypoint=701
 
 ; Determines how many pixels should the editor looks outside of the screen for objects to be drawn.
 ; Some padding is necessary so large objects whose origin cell is just outside of the screen
@@ -27,16 +27,16 @@ RenderPixelPadding=50
 ; Determines whether the world is flat.
 ; If set, disables some height-specific functionality
 ; that is not necessary when height levels are not used.
-IsFlatWorld=true
+IsFlatWorld=false
 
 ; Does Tiberium use the Theater palette instead of the Unit palette?
 TheaterPaletteForTiberium=true
 
 ; Do Veins use the Theater palette instead of the Unit palette?
-TheaterPaletteForVeins=false
+TheaterPaletteForVeins=true
 
 ; Should Tiberium be affected by map lighting?
-TiberiumAffectedByLighting=true
+TiberiumAffectedByLighting=false
 
 ; Should Tiberium spawning terrain objects be affected by map lighting?
 TiberiumTreesAffectedByLighting=false
@@ -50,7 +50,7 @@ VoxelsAffectedByLighting=false
 ; The file name(s) of the executable file(s) that the map editor tries to find from the game directory.
 ; Used for the verification that the user has given us the correct game directory.
 ; You can input multiple executable names by separating them with a comma (,).
-ExpectedClientExecutableName=DTA.exe
+ExpectedClientExecutableName=ra2md.exe
 
 ; Specifies the path(s) that we should check in the Windows registry to determine
 ; the game installation directory (if one hasn't been defined in the map editor settings file).
@@ -59,16 +59,16 @@ ExpectedClientExecutableName=DTA.exe
 ; By default, the registry key is expected to exist in the current user's registry (HKEY_CURRENT_USER).
 ; If you begin a path with "HKLM:", the editor expects to find the path from the
 ; local machine registry (HKEY_LOCAL_MACHINE) instead.
-GameRegistryInstallPath=SOFTWARE\DawnOfTheTiberiumAge
+GameRegistryInstallPath=SOFTWARE\YurisRevenge,HKLM:SOFTWARE\WOW6432Node\Westwood\Yuri's Revenge
 
 ; If set to yes, WAE will always write a dummy preview to a map file that does not have a preview.
 ; This is required by vanilla games (both TS as well as RA2/YR) as they crash if a map has no preview.
-DefaultPreview=false
+DefaultPreview=true
 
 ; Specifies the file filter string used in OpenFileDialog (Windows)
 ; For multiple file formats, list them with : (instead of ;)
 ; Example config with multiple file extensions: "YR maps|*.map:*.mpr:*.yrm|All files|*.*"
-OpenFileDialogFilter=TS maps|*.map|All files|*.*
+OpenFileDialogFilter=All RA2/YR maps|*.map:*.mpr:*.yrm|RA2/YR singleplayer and CnCNet multiplayer maps|*.map|Non-CnCNet RA2 multiplayer maps|*.mpr|Non-CnCNet YR multiplayer maps|*.yrm|All files|*.*
 
 ; Should the editor consider Ares [#include] or Phobos [$Include] section?
 EnableIniInclude=false
@@ -83,11 +83,11 @@ IntegerVariables=false
 AdvancedFacingsHack=true
 
 ; Is NewTheater generic fallback (letter G) available?
-NewTheaterGenericBuilding=false
+NewTheaterGenericBuilding=true
 
 ; Should the editor assume that it is used for RA2/YR, as opposed to TS?
 ; This includes, but is not limited to reading and using RA2/YR Countries rather than TS HouseTypes.
-IsRA2YR=false
+IsRA2YR=true
 
 ; Maximum allowed dimensions of generated map previews.
 MapPreviewMaxWidth=400
@@ -95,13 +95,13 @@ MapPreviewMaxHeight=300
 
 ; Paths to various files that the editor loads
 [FilePaths]
-Rules=INI/Rules.ini
-Firestorm=INI/Enhance.ini
-Art=INI/Art.ini
-ArtFS=INI/ArtE.ini
-AI=INI/AI.ini
-AIE=INI/AIE.ini
-Tutorial=INI/Tutorial.ini
-Theme=INI/Theme.ini
-Sound=INI/Sound01.ini
-EVA=
+Rules=rulesmd.ini
+Firestorm=
+Art=artmd.ini
+ArtFS=
+AI=aimd.ini
+AIFS=
+Tutorial=
+Theme=thememd.ini
+Sound=soundmd.ini
+EVA=evamd.ini
