+++ src/TSMapEditor/Config/Default/FileManagerConfig.ini
@@ -1,9 +1,9 @@
-﻿; C&C World-Altering Editor (WAE)
+; C&C World-Altering Editor (WAE)
 ; https://github.com/CnCNet/WorldAlteringEditor
 
 ; Specifies directories that the file manager will load MIX files from.
 [SearchDirectories]
-0=MIX
+0=.
 1=Map Editor
 
 ; Specifies the MIX files loaded by the editor.
@@ -13,12 +13,26 @@
 ; Special values are supported to load additional MIX files. These are always optional.
 ; The valid options are $TSECACHE, $RA2ECACHE, $TSELOCAL, $RA2ELOCAL, $EXPAND, $EXPANDMD.
 [MIXFiles]
-0=DTACache.mix,0
-1=$TSECACHE
-2=Conquer.mix,1
-3=Local.mix,1
-4=Cache.mix,1
-5=Sidec01.mix,0
-6=Sidec02.mix,0
-7=Sidec03.mix,0
-8=Sidec04.mix,0
+0=WAECache.mix
+1=langmd.mix,1
+2=language.mix,1
+3=$EXPANDMD
+4=ra2md.mix,1
+5=ra2.mix,1
+6=cachemd.mix,1
+7=cache.mix,1
+8=localmd.mix,1
+9=local.mix,1
+10=$RA2ECACHE
+11=$RA2ELOCAL
+12=conqmd.mix,1
+13=genermd.mix,1
+14=generic.mix,1
+15=isogenmd.mix,1
+16=isogen.mix,1
+17=conquer.mix,1
+18=marble.mix,0
+
+[StringTables]
+0=ra2.csf
+1=ra2md.csf
