+++ src/TSMapEditor/Config/Default/OverlayCollections.ini
@@ -4,87 +4,63 @@
 ; Specifies overlay collections available in the sidebar.
 
 [OverlayCollections]
-0=GreenTiberium
-1=BlueTiberium
-2=Ore
-3=ScrapMetal
-4=GreenGems
-5=BlueGems
-6=MixedGems
-7=AutumnGrass
-
-[GreenTiberium]
-Name=Green Tiberium
-OverlayType0=RTIB01
-OverlayType1=RTIB02
-OverlayType2=RTIB03
-OverlayType3=RTIB04
-OverlayType4=RTIB05
-OverlayType5=RTIB06
-OverlayType6=RTIB07
-OverlayType7=RTIB08
-OverlayType8=RTIB09
-OverlayType9=RTIB10
-OverlayType10=RTIB11
-OverlayType11=RTIB12
-
-[BlueTiberium]
-Name=Blue Tiberium
-OverlayType0=QTIB01
-OverlayType1=QTIB02
-OverlayType2=QTIB03
-OverlayType3=QTIB04
-OverlayType4=QTIB05
-OverlayType5=QTIB06
-OverlayType6=QTIB07
-OverlayType7=QTIB08
-OverlayType8=QTIB09
-OverlayType9=QTIB10
-OverlayType10=QTIB11
-OverlayType11=QTIB12
+0=Ore
+1=Gems
+2=ClearRocks
+3=SandRocks
+4=LunarRocks
 
 [Ore]
 Name=Ore
-OverlayType0=ZORE01
-OverlayType1=ZORE02
-OverlayType2=ZORE03
-OverlayType3=ZORE04
+OverlayType0=TIB01
+OverlayType1=TIB02
+OverlayType2=TIB03
+OverlayType3=TIB04
+OverlayType4=TIB05
+OverlayType5=TIB06
+OverlayType6=TIB07
+OverlayType7=TIB08
+OverlayType8=TIB09
+OverlayType9=TIB10
+OverlayType10=TIB11
+OverlayType11=TIB12
 
-[ScrapMetal]
-Name=Scrap Metal
-OverlayType0=SCRAP01
-OverlayType1=SCRAP02
-OverlayType2=SCRAP03
-OverlayType3=SCRAP04
+[Gems]
+Name=Gems
+OverlayType0=GEM01
+OverlayType1=GEM02
+OverlayType2=GEM03
+OverlayType3=GEM04
+OverlayType4=GEM05
+OverlayType5=GEM06
+OverlayType6=GEM07
+OverlayType7=GEM08
+OverlayType8=GEM09
+OverlayType9=GEM10
+OverlayType10=GEM11
+OverlayType11=GEM12
 
-[GreenGems]
-Name=Green Gems
-OverlayType0=GRNGEM01
-OverlayType1=GRNGEM02
-OverlayType2=GRNGEM03
-OverlayType3=GRNGEM04
+[ClearRocks]
+Name=Clear Rocks
+OverlayType0=TROCK01
+OverlayType1=TROCK02
+OverlayType2=TROCK03
+OverlayType3=TROCK04
+OverlayType4=TROCK05
 
-[BlueGems]
-Name=Blue Gems
-OverlayType0=BLUGEM01
-OverlayType1=BLUGEM02
-OverlayType2=BLUGEM03
-OverlayType3=BLUGEM04
+[SandRocks]
+Name=Sand Rocks
+OverlayType0=SROCK01
+OverlayType1=SROCK02
+OverlayType2=SROCK03
+OverlayType3=SROCK04
+OverlayType4=SROCK05
 
-[MixedGems]
-Name=Mixed Gems
-OverlayType0=MIXGEM01
-OverlayType1=MIXGEM02
-OverlayType2=MIXGEM03
-OverlayType3=MIXGEM04
-
-; Demonstrates including optional overlay frames in the types of overlay to place.
-[AutumnGrass]
-Name=Small Autumn Grass
-OverlayType0=AJGRASS,0
-OverlayType1=AJGRASS,1
-OverlayType2=AJGRASS,2
-OverlayType3=AJGRASS,3
-OverlayType4=AJGRASS,4
-OverlayType5=AJGRASS,5
-
+[LunarRocks]
+Name=Lunar Rocks
+OverlayType0=LUNRK1
+OverlayType1=LUNRK2
+OverlayType2=LUNRK3
+OverlayType3=LUNRK4
+OverlayType4=LUNRK5
+AllowedTheaters=LUNAR
