+++ src/TSMapEditor/Config/Default/ScriptActions.ini
@@ -1,4 +1,4 @@
-﻿; C&C World-Altering Editor (WAE)
+; C&C World-Altering Editor (WAE)
 ; https://github.com/CnCNet/WorldAlteringEditor
 
 ; This file defines the script actions available in the map editor.
@@ -25,7 +25,8 @@ Option6=6,Factories
 Option7=7,Base Defenses
 Option8=8,Base Threats
 Option9=9,Power Plants
-Option10=10,Harvesters (Vinifera)
+Option10=10,Occupiables
+Option11=11,Tech Buildings
 
 [AttackWaypoint]
 Name=Attack Waypoint
@@ -95,22 +96,29 @@ Option5=5,Guard
 Option6=6,Sticky
 Option7=7,Enter
 Option8=8,Capture
-Option9=9,Harvest
-Option10=10,Area Guard
-Option11=11,Return
-Option12=12,Stop
-Option13=13,Ambush
-Option14=14,Hunt
-Option15=15,Unload
-Option16=16,Sabotage
-Option17=17,Construction
-Option18=18,Selling
-Option19=19,Repair
-Option20=20,Rescue
-Option21=21,Missile
-Option22=22,Harmless
-Option23=23,Open
-Option24=24,Patrol
+Option9=9,Eaten
+Option10=10,Harvest
+Option11=11,Area Guard
+Option12=12,Return
+Option13=13,Stop
+Option14=14,Ambush
+Option15=15,Hunt
+Option16=16,Unload
+Option17=17,Sabotage
+Option18=18,Construction
+Option19=19,Selling
+Option20=20,Repair
+Option21=21,Rescue
+Option22=22,Missile
+Option23=23,Harmless
+Option24=24,Open
+Option25=25,Patrol
+Option26=26,Paradrop Approach
+Option27=27,Paradrop Overfly
+Option28=28,Wait
+Option29=29,Attack (Again)
+Option30=30,Spyplane Approach
+Option31=31,Spyplane Overfly
 
 [SetGlobalVariable]
 Name=Set Global Variable
@@ -337,3 +345,112 @@ Option0=0,None
 Option1=1,Asterisk (*)
 Option2=2,Question mark (?)
 Option3=3,Exclamation mark (!)
+
+; ********** RA2/YR BEGINS + ARES ALTERATIONS ***********
+[GatherAtEnemy]
+Name=Gather at Enemy (Ares Improved)
+Description=Gather on a spot near enemy base at a distance in cells given by AISafeDistance + Number (+ve or -ve). Ares makes use of the unused Number parameter. If not using Ares for YR then set the parameter to 0.
+ParamType=Number
+ParamDescription=Distance Variable
+
+[GatherAtBase]
+Name=Gather at Base (Ares Improved)
+DescriptionGathers outside own base perimeter at a distance in cells given by AISafeDistance (Ares: AIFriendlyDistance) + Number (+ve or -ve). Ares makes use of the unused Number parameter. If not using Ares for YR then set the parameter to 0.
+ParamType=Number
+ParamDescription=Distance Variable
+
+[IronCurtain]
+Name=Iron Curtain (Ares Improved)
+Description=Calls (and waits if nearly ready) for the House to deliver Iron Curtain of the specified superweapon group Number to team. Ares makes use of the unused Number parameter. If not using Ares for YR then set the parameter to 0.
+ParamType=Number; a superweapon selector would be useful.
+ParamDescription=SW.Group ID
+
+[ChronoSphereTeamBuildings]
+Name=ChronoSphere Team to Building
+Description=If the owner of the team has the Chronosphere fully charged then it will be used to move the team to the given structure number before moving on to next action. 
+ParamDescription=Building
+ParamType=BuildingWithProperty
+
+[ChronoSphereTeamTarget]
+Name=ChronoSphere Team to target Type
+Description=If the owner of the team has the Chronosphere fully charged then it will be used to transport the task force to the given target type. Exact target is either nearest/first built based on the value used in TargetDistanceCoefficientDefault or its technotype override of TargetDistanceCoefficient.
+ParamDescription=Target
+ParamType=Quarry
+Option0=0,N/A (Cancel)
+Option1=1,Anything
+Option2=2,Structures
+Option3=3,Harvesters
+Option4=4,Infantry
+Option5=5,Vehicles
+Option6=6,Factories
+Option7=7,Base Defenses
+Option8=8,Base Threats
+Option9=9,Power Plants
+Option10=10,Occupiables
+Option11=11,Tech Buildings
+
+[MoveToFriendlyStructure]
+Name=Move to Friendly Structure
+Description=16,0,1,Move to own structure or owned by allies given in the structure number parameter.
+ParamDescription=Building
+ParamType=BuildingWithProperty
+
+[AttackStructureAtWaypoint]
+Name=Attack structure at waypoint
+Description=Attacks the structure at given waypoint. Owner is not considered as of enemy or ally or neutral. An attack by Engineer would mean repairing the building. This action also uses the weapon if present of the unit instead its special abilities like C4. This is useful in cases like destroying an occupiable structure instead of entering it.
+ParamType=Waypoint
+ParamDescription=Waypoint ID
+
+[EnterGrinder]
+Name=Enter Grinder
+Description=Team will enter nearest grinder owned by the owner of this team or allies.
+
+[OccupyTankBunker]
+Name=Occupy Tank Bunker
+Description=Any vehicle members of this team which do not have Bunkerable=no set will enter the nearest vacant structure with Bunker=yes set if it is owned by the owner of the team or any of its allies.
+
+[EnterBioReactor]
+Name=Enter bio reactor
+Description=Any infantry members of this team will enter the nearest structure with InfantryAbsorb=yes set if it is owned by the owner of this team or any of its allies and has space available.
+
+[OccupyBattleBunker];63
+Name=Occupy Battle Bunker
+Description=Any infantry members of this team with Occupier=yes set will enter the nearest structure (which they own) with CanOccupyFire=yes set.
+
+[GarrisonStructure]
+Name=Garrison Structure
+Description=Any infantry members of this team with Occupier=yes set will enter the nearest (neutral) structure with CanOccupyFire=yes set. It is specifically used so as to transfer the unit's weapon to the structure.
+
+; ********** ARES ADDITIONS *********** Starts at 65
+
+[AuxPower]
+Name=Auxiliary Power [Ares]
+Description=Permanently changes the power output of the house owning the team by the specified amount. 
+ParamType=Number
+ParamDescription=Power Amount
+
+[KillDrivers]
+Name=Kill Drivers [Ares]
+Description=Kills all unprotected drivers of the units in this team and assigns them to the country called Special.
+
+[TakeVehicles]
+Name=Take Vehicles [Ares]
+Description=If possible, all infantry (CanDrive or VehicleThief) in this team will be assigned to the closest vehicle they can drive or hijack.
+
+[ConvertType]
+Name=Convert Type [Ares]
+Description=Changes all members of this team into their respective script conversion types defined by their respective Ares Convert.Script flag.
+
+[SonarReveal]
+Name=Sonar Reveal [Ares]
+Description=Disables the ability of all team members to cloak themselves for the given Duration (number of frames).
+ParamType=Number
+ParamDescription=Frame Duration
+
+[DisableWeapons]
+Name=Disable Weapons [Ares]
+Description=Disables the ability of all team members to fire for the given Duration (number of frames). Use 0 to end the effect.
+ParamType=Number
+ParamDescription=Frame Duration
+
+; ********** ARES ENDS ***********
