+++ src/TSMapEditor/Config/Default/SmudgeCollections.ini
@@ -5,92 +5,56 @@
 
 [SmudgeCollections]
 0=Craters
-1=Burns
-2=SnowyCraters
-3=SnowyBurns
+1=SmallCraters
+2=LargeCraters
+3=Burns
 
 [Craters]
 Name=Random Crater
-SmudgeType0=CR1E ; first entry is used for the texture on the sidebar,
-                 ; so we want it to use a big crater
-SmudgeType1=CR1B
-SmudgeType2=CR1C
-SmudgeType3=CR1D
-SmudgeType4=CR1A
-SmudgeType5=CR2A
-SmudgeType6=CR2B
-SmudgeType7=CR2C
-SmudgeType8=CR2D
-SmudgeType9=CR2E
-SmudgeType10=CR3A
-SmudgeType11=CR3B
-SmudgeType12=CR3C
-SmudgeType13=CR3D
-SmudgeType14=CR3E
-SmudgeType15=CR4A
-SmudgeType16=CR4B
-SmudgeType17=CR4C
-SmudgeType18=CR4D
-SmudgeType19=CR4E
-SmudgeType20=CR5A
-SmudgeType21=CR5B
-SmudgeType22=CR5C
-SmudgeType23=CR5D
-SmudgeType24=CR5E
-SmudgeType25=CR6A
-SmudgeType26=CR6B
-SmudgeType27=CR6C
-SmudgeType28=CR6D
-SmudgeType29=CR6E
+SmudgeType0=CRATER12 ; first entry is used for the texture on the sidebar,
+                     ; so we want it to use a big crater
+SmudgeType1=CRATER11
+SmudgeType2=CRATER10
+SmudgeType3=CRATER09
+SmudgeType4=CRATER08
+SmudgeType5=CRATER07
+SmudgeType6=CRATER06
+SmudgeType7=CRATER05
+SmudgeType8=CRATER04
+SmudgeType9=CRATER03
+SmudgeType10=CRATER02
+SmudgeType11=CRATER01
+
+[SmallCraters]
+Name=Random 1x1 Crater
+SmudgeType0=CRATER10
+SmudgeType1=CRATER09
+SmudgeType2=CRATER08
+SmudgeType3=CRATER07
+SmudgeType4=CRATER06
+SmudgeType5=CRATER05
+SmudgeType6=CRATER04
+SmudgeType7=CRATER03
+SmudgeType8=CRATER02
+SmudgeType9=CRATER01
+
+[LargeCraters]
+Name=Random 2x2 Crater
+SmudgeType0=CRATER12
+SmudgeType1=CRATER11
 
 [Burns]
 Name=Random Burn
-SmudgeType0=SC1
-SmudgeType1=SC2
-SmudgeType2=SC3
-SmudgeType3=SC4
-SmudgeType4=SC5
-SmudgeType5=SC6
-
-[SnowyCraters]
-Name=Random Snowy Crater
-SmudgeType0=CRS1E ; first entry is used for the texture on the sidebar,
-                  ; so we want it to use a big crater
-SmudgeType1=CRS1B
-SmudgeType2=CRS1C
-SmudgeType3=CRS1D
-SmudgeType4=CRS1A
-SmudgeType5=CRS2A
-SmudgeType6=CRS2B
-SmudgeType7=CRS2C
-SmudgeType8=CRS2D
-SmudgeType9=CRS2E
-SmudgeType10=CRS3A
-SmudgeType11=CRS3B
-SmudgeType12=CRS3C
-SmudgeType13=CRS3D
-SmudgeType14=CRS3E
-SmudgeType15=CRS4A
-SmudgeType16=CRS4B
-SmudgeType17=CRS4C
-SmudgeType18=CRS4D
-SmudgeType19=CRS4E
-SmudgeType20=CRS5A
-SmudgeType21=CRS5B
-SmudgeType22=CRS5C
-SmudgeType23=CRS5D
-SmudgeType24=CRS5E
-SmudgeType25=CRS6A
-SmudgeType26=CRS6B
-SmudgeType27=CRS6C
-SmudgeType28=CRS6D
-SmudgeType29=CRS6E
-
-[SnowyBurns]
-Name=Random Snowy Burn
-SmudgeType0=SCS1
-SmudgeType1=SCS2
-SmudgeType2=SCS3
-SmudgeType3=SCS4
-SmudgeType4=SCS5
-SmudgeType5=SCS6
+SmudgeType0=BURNT12 ; first entry is used for the texture on the sidebar,
+                     ; so we want it to use a big crater
+SmudgeType1=BURNT11
+SmudgeType2=BURNT10
+SmudgeType3=BURNT09
+SmudgeType4=BURNT08
+SmudgeType5=BURNT07
+SmudgeType6=BURNT06
+SmudgeType7=BURNT05
+SmudgeType8=BURNT04
+SmudgeType9=BURNT03
+SmudgeType10=BURNT02
+SmudgeType11=BURNT01
