+++ src/TSMapEditor/Config/Default/TeamTypeFlags.ini
@@ -8,6 +8,7 @@
 Full=false
 Whiner=false
 Droppod=false
+UseTransportOrigin=false
 Suicide=false
 Loadable=false
 Prebuild=false
@@ -25,4 +26,3 @@ IsBaseDefense=false
 OnlyTargetHouseEnemy=false
 TransportsReturnOnUnload=false
 AreTeamMembersRecruitable=false
-SmartHunt=true
