+++ src/TSMapEditor/Config/Default/TerrainObjectCollections.ini
@@ -1,88 +1,91 @@
-﻿; C&C World-Altering Editor (WAE)
+; C&C World-Altering Editor (WAE)
 ; https://github.com/CnCNet/WorldAlteringEditor
 
 ; Specifies terrain object collections available in the sidebar.
 
 [TerrainObjectCollections]
-0=TemperateConifers
-1=LeafyTrees
-2=AutumnTrees
-3=AltTerrainConifers
-4=AltTerrainLeafyTrees
-5=BurnedConifers
-6=BurnedLeafyTrees
+0=DeciduousTrees
+1=TropicalTrees
+2=ConiferTrees
+3=SnowConiferTrees
+4=TemperateCactus
+5=Cacti
+6=ExoticTrees
+7=DeadTrees
 
-[TemperateConifers]
-Name=Random Temperate Conifer
-TerrainObjectType0=T01
-TerrainObjectType1=T02
-TerrainObjectType2=T05
-TerrainObjectType3=T06
-TerrainObjectType4=T07
-TerrainObjectType5=T08
-TerrainObjectType6=T09
-TerrainObjectType7=T16
+[DeciduousTrees]
+Name=Deciduous Trees
+TerrainObjectType0=TREE01
+TerrainObjectType1=TREE02
+TerrainObjectType2=TREE03
+TerrainObjectType3=TREE04
+TerrainObjectType4=TREE05
+TerrainObjectType5=TREE06
+TerrainObjectType6=TREE07
+TerrainObjectType7=TREE08
+TerrainObjectType8=TREE09
+TerrainObjectType9=TREE10
+TerrainObjectType10=TREE11
+TerrainObjectType11=TREE12
+TerrainObjectType12=TREE13
+TerrainObjectType13=TREE14
+TerrainObjectType14=TREE15
+TerrainObjectType15=TREE16
+TerrainObjectType16=TREE17
+TerrainObjectType17=TREE18
+TerrainObjectType18=TREE19
+AllowedTheaters=TEMPERATE,SNOW,URBAN,NEWURBAN,DESERT
 
-[LeafyTrees]
-Name=Random Leafy Tree
-TerrainObjectType0=T03
-TerrainObjectType1=T10
-TerrainObjectType2=T11
-TerrainObjectType3=T12
-TerrainObjectType4=T13
-TerrainObjectType5=T14
-TerrainObjectType6=T15
-TerrainObjectType7=T17
+[TropicalTrees]
+Name=Tropical Trees
+TerrainObjectType0=TREE20
+TerrainObjectType1=TREE21
+TerrainObjectType2=TREE22
+TerrainObjectType3=TREE23
+TerrainObjectType4=TREE24
+AllowedTheaters=TEMPERATE,URBAN,NEWURBAN,DESERT
 
-[AutumnTrees]
-Name=Random Autumn Tree
-TerrainObjectType0=AT03
-TerrainObjectType1=AT10
-TerrainObjectType2=AT11
-TerrainObjectType3=AT12
-TerrainObjectType4=AT13
-TerrainObjectType5=AT14
-TerrainObjectType6=AT15
-TerrainObjectType7=AT17
+[ConiferTrees]
+Name=Conifer Trees
+TerrainObjectType0=TREE25
+TerrainObjectType1=TREE26
+TerrainObjectType2=TREE27
+AllowedTheaters=TEMPERATE,URBAN,NEWURBAN,DESERT
 
-[AltTerrainConifers]
-Name=Random Alt. Conifer
-TerrainObjectType0=ST01
-TerrainObjectType1=ST02
-TerrainObjectType2=ST05
-TerrainObjectType3=ST06
-TerrainObjectType4=ST07
-TerrainObjectType5=ST08
-TerrainObjectType6=ST16
+[SnowConiferTrees]
+Name=Conifer Trees
+TerrainObjectType0=TREE20
+TerrainObjectType1=TREE21
+TerrainObjectType2=TREE22
+TerrainObjectType3=TREE23
+TerrainObjectType4=TREE24
+TerrainObjectType5=TREE25
+TerrainObjectType6=TREE26
+TerrainObjectType7=TREE27
+AllowedTheaters=SNOW
 
-[AltTerrainLeafyTrees]
-Name=Random Alt. Leafy Tree
-TerrainObjectType0=ST03
-TerrainObjectType1=ST10
-TerrainObjectType2=ST11
-TerrainObjectType3=ST12
-TerrainObjectType4=ST13
-TerrainObjectType5=ST14
-TerrainObjectType6=ST15
-TerrainObjectType7=ST17
+[TemperateCactus]
+Name=Cacti
+TerrainObjectType0=TREE28
+AllowedTheaters=TEMPERATE,URBAN
 
-[BurnedConifers]
-Name=Random Burned Conifer
-TerrainObjectType0=BT01
-TerrainObjectType1=BT02
-TerrainObjectType2=BT05
-TerrainObjectType3=BT06
-TerrainObjectType4=BT07
-TerrainObjectType5=BT16
+[Cacti]
+Name=Cacti
+TerrainObjectType0=TREE28
+TerrainObjectType1=TREE29
+TerrainObjectType2=TREE30
+AllowedTheaters=NEWURBAN,DESERT
 
-[BurnedLeafyTrees]
-Name=Random Burned Leafy Tree
-TerrainObjectType0=BT03
-TerrainObjectType1=BT10
-TerrainObjectType2=BT11
-TerrainObjectType3=BT12
-TerrainObjectType4=BT13
-TerrainObjectType5=BT14
-TerrainObjectType6=BT15
-TerrainObjectType7=BT17
+[ExoticTrees]
+Name=Exotic Trees
+TerrainObjectType0=TREE31
+TerrainObjectType1=TREE32
+TerrainObjectType2=TREE33
+AllowedTheaters=NEWURBAN,DESERT
 
+[DeadTrees]
+Name=Dead Trees
+TerrainObjectType0=TREE34
+TerrainObjectType1=TREE35
+TerrainObjectType2=TREE36
+AllowedTheaters=NEWURBAN,DESERT
