+++ src/TSMapEditor/Config/Default/TerrainTypeImpassability.ini
@@ -3,128 +3,110 @@
 
 ; Specifies which cells of terrain objects should be marked as impassable.
 
-; ******************
-; Autumn Tree Groups
-; ******************
-
-[ATC02]
-ImpassableCells=0,0|0,1|1,0|1,1
-
-[ATC05]
-ImpassableCells=0,0|0,-1|-1,0|1,0|1,1
-
-; ******
-; Bushes
-; ******
-
-[BR6]
-ImpassableCells=0,0|0,1|1,0|1,1
-
-[BR7]
-ImpassableCells=0,0|1,0|1,1
-
-[BR8]
-ImpassableCells=0,0|1,0|2,0|1,1
-
-[BR9]
-ImpassableCells=0,0|0,1|1,0|1,1
-
-; *************
-; Autumn Bushes
-; *************
-
-[ABR6]
-ImpassableCells=0,0|0,1|1,0|1,1
-
-[ABR7]
-ImpassableCells=0,0|1,0|1,1
-
-[ABR8]
-ImpassableCells=0,0|1,0|2,0|1,1
-
-[ABR9]
-ImpassableCells=0,0|0,1|1,0|1,1
-
-; ***************
-; Alt Tree Groups
-; ***************
-
-[STC01]
-ImpassableCells=0,0
-
-[STC02]
-ImpassableCells=0,0|0,1|1,0|1,1
-
-[STC03]
-ImpassableCells=0,0|0,1|1,0|1,1
-
-[STC04]
-ImpassableCells=0,0|0,-1|-1,0
-
-[STC05]
-ImpassableCells=0,0|0,-1|-1,0|1,0|1,1
-
-; ******************
-; Burned Tree Groups
-; ******************
-
-[BTC01]
-ImpassableCells=0,0
-
-[BTC02]
-ImpassableCells=0,0|0,1|1,0|1,1
-
-[BTC03]
-ImpassableCells=0,0|0,1|1,0|1,1
-
-[BTC04]
-ImpassableCells=0,0|0,-1|-1,0
-
-[BTC05]
-ImpassableCells=0,0|0,-1|-1,0|1,0|1,1
-
-; *******************
-; Regular Tree Groups
-; *******************
-
-[TC01]
-ImpassableCells=0,0
-
-[TC02]
-ImpassableCells=0,0|0,1|1,0|1,1
-
-[TC03]
-ImpassableCells=0,0|0,1|1,0|1,1
-
-[TC04]
-ImpassableCells=0,0|0,-1|-1,0
-
-[TC05]
-ImpassableCells=0,0|0,-1|-1,0|1,0|1,1
-
-; **********
-; Sand Rocks
-; **********
-
-[SROCK01]
-ImpassableCells=0,0
-
-[SROCK02]
-ImpassableCells=0,0
-
-[SROCK03]
-ImpassableCells=0,0
-
-[SROCK04]
-ImpassableCells=0,0
-
-[SROCK05]
-ImpassableCells=0,0
-
-[SROCK06]
-ImpassableCells=0,0|1,0|0,1
-
-[SROCK07]
-ImpassableCells=0,0|1,0|0,1
-
-
+; ; ******************
+; ; Autumn Tree Groups
+; ; ******************
+; 
+; [ATC02]
+; ImpassableCells=0,0|0,1|1,0|1,1
+; 
+; [ATC05]
+; ImpassableCells=0,0|0,-1|-1,0|1,0|1,1
+; 
+; ; ******
+; ; Bushes
+; ; ******
+; 
+; [BR6]
+; ImpassableCells=0,0|0,1|1,0|1,1
+; 
+; [BR7]
+; ImpassableCells=0,0|1,0|1,1
+; 
+; [BR8]
+; ImpassableCells=0,0|1,0|2,0|1,1
+; 
+; [BR9]
+; ImpassableCells=0,0|0,1|1,0|1,1
+; 
+; ; ***************
+; ; Alt Tree Groups
+; ; ***************
+; 
+; [STC01]
+; ImpassableCells=0,0
+; 
+; [STC02]
+; ImpassableCells=0,0|0,1|1,0|1,1
+; 
+; [STC03]
+; ImpassableCells=0,0|0,1|1,0|1,1
+; 
+; [STC04]
+; ImpassableCells=0,0|0,-1|-1,0
+; 
+; [STC05]
+; ImpassableCells=0,0|0,-1|-1,0|1,0|1,1
+; 
+; ; ******************
+; ; Burned Tree Groups
+; ; ******************
+; 
+; [BTC01]
+; ImpassableCells=0,0
+; 
+; [BTC02]
+; ImpassableCells=0,0|0,1|1,0|1,1
+; 
+; [BTC03]
+; ImpassableCells=0,0|0,1|1,0|1,1
+; 
+; [BTC04]
+; ImpassableCells=0,0|0,-1|-1,0
+; 
+; [BTC05]
+; ImpassableCells=0,0|0,-1|-1,0|1,0|1,1
+; 
+; ; *******************
+; ; Regular Tree Groups
+; ; *******************
+; 
+; [TC01]
+; ImpassableCells=0,0
+; 
+; [TC02]
+; ImpassableCells=0,0|0,1|1,0|1,1
+; 
+; [TC03]
+; ImpassableCells=0,0|0,1|1,0|1,1
+; 
+; [TC04]
+; ImpassableCells=0,0|0,-1|-1,0
+; 
+; [TC05]
+; ImpassableCells=0,0|0,-1|-1,0|1,0|1,1
+; 
+; ; **********
+; ; Sand Rocks
+; ; **********
+; 
+; [SROCK01]
+; ImpassableCells=0,0
+; 
+; [SROCK02]
+; ImpassableCells=0,0
+; 
+; [SROCK03]
+; ImpassableCells=0,0
+; 
+; [SROCK04]
+; ImpassableCells=0,0
+; 
+; [SROCK05]
+; ImpassableCells=0,0
+; 
+; [SROCK06]
+; ImpassableCells=0,0|1,0|0,1
+; 
+; [SROCK07]
+; ImpassableCells=0,0|1,0|0,1
