+++ src/TSMapEditor/Config/Default/Theaters.ini
@@ -5,24 +5,72 @@
 
 
 [Theaters]
-0=Temperate
-1=Desert
+0=TEMPERATE
+1=SNOW
+2=DESERT
+3=URBAN
+4=NEWURBAN
+5=LUNAR
 
-[Temperate]
-ConfigINIPath=INI/Temperat.ini
-ContentMIXName=IsoTemp.mix
-OptionalContentMIXName=Temperat.mix
+[TEMPERATE]
+ConfigINIPath=temperatmd.ini
+ContentMIXName=isotemmd.mix,isotemp.mix,temperat.mix,tem.mix
 TerrainPaletteName=isotem.pal
 UnitPaletteName=unittem.pal
-;TiberiumPaletteName=
+TiberiumPaletteName=temperat.pal
 FileExtension=.tem
-NewTheaterBuildingLetter=A
+NewTheaterBuildingLetter=T
+PaveConnectToTileSets=20,38,40
+GreenConnectToTileSets=12
 
-[Desert]
-ConfigINIPath=INI/Desert.ini
-ContentMIXName=IsoDes.mix
-OptionalContentMIXName=Desert.mix
+[SNOW]
+ConfigINIPath=snowmd.ini
+ContentMIXName=isosnomd.mix,isosnow.mix,snowmd.mix,snow.mix,sno.mix
+TerrainPaletteName=isosno.pal
+UnitPaletteName=unitsno.pal
+TiberiumPaletteName=snow.pal
+FileExtension=.sno
+NewTheaterBuildingLetter=A
+PaveConnectToTileSets=20,38,40
+GreenConnectToTileSets=12
+
+[DESERT]
+ConfigINIPath=desertmd.ini
+ContentMIXName=isodesmd.mix,isodes.mix,desert.mix,des.mix
 TerrainPaletteName=isodes.pal
 UnitPaletteName=unitdes.pal
+TiberiumPaletteName=desert.pal
 FileExtension=.des
-NewTheaterBuildingLetter=T
+NewTheaterBuildingLetter=D
+PaveConnectToTileSets=20,38,40
+GreenConnectToTileSets=12
+
+[URBAN]
+ConfigINIPath=urbanmd.ini
+ContentMIXName=isourbmd.mix,isourb.mix,urban.mix,urb.mix
+TerrainPaletteName=isourb.pal
+UnitPaletteName=uniturb.pal
+TiberiumPaletteName=urban.pal
+FileExtension=.urb
+NewTheaterBuildingLetter=U
+
+[NEWURBAN]
+ConfigINIPath=urbannmd.ini
+ContentMIXName=isoubnmd.mix,isoubn.mix,urbann.mix,ubn.mix
+TerrainPaletteName=isoubn.pal
+UnitPaletteName=unitubn.pal
+TiberiumPaletteName=urbann.pal
+FileExtension=.ubn
+FallbackTileFileExtension=.urb ; for loading .urb files instead of .ubn files from Marble.mix
+NewTheaterBuildingLetter=N
+PaveConnectToTileSets=20,38,40
+GreenConnectToTileSets=12
+
+[LUNAR]
+ConfigINIPath=lunarmd.ini
+ContentMIXName=isolunmd.mix,isolun.mix,lunar.mix,lun.mix
+TerrainPaletteName=isolun.pal
+UnitPaletteName=unitlun.pal
+TiberiumPaletteName=lunar.pal
+FileExtension=.lun
+NewTheaterBuildingLetter=L
