+++ src/TSMapEditor/Config/Default/Tiberiums.ini
@@ -4,14 +4,7 @@
 ; A list of overlays used by each Tiberium type.
 
 [Tiberiums]
-Riparius=RTIB01,RTIB02,RTIB03,RTIB04,RTIB05,RTIB06,RTIB07,RTIB08,RTIB09,RTIB10,RTIB11,RTIB12
-Gems=MIXGEM01,MIXGEM02,MIXGEM03,MIXGEM04
-Vinifera=QTIB01,QTIB02,QTIB03,QTIB04,QTIB05,QTIB06,QTIB07,QTIB08,QTIB09,QTIB10,QTIB11,QTIB12
-Ore=ZORE01,ZORE02,ZORE03,ZORE04
-Scrap=SCRAP01,SCRAP02,SCRAP03,SCRAP04
-GreenGems=GRNGEM01,GRNGEM02,GRNGEM03,GRNGEM04
-BlueGems=BLUGEM01,BLUGEM02,BLUGEM03,BLUGEM04
-SnowyRiparius=TIB1S_01,TIB1S_02,TIB1S_03,TIB1S_04,TIB1S_05,TIB1S_06,TIB1S_07,TIB1S_08,TIB1S_09,TIB1S_10,TIB1S_11,TIB1S_12
-SnowyVinifera=TIB2S_01,TIB2S_02,TIB2S_03,TIB2S_04,TIB2S_05,TIB2S_06,TIB2S_07,TIB2S_08,TIB2S_09,TIB2S_10,TIB2S_11,TIB2S_12
-SnowyOre=ORESNO01,ORESNO02,ORESNO03,ORESNO04
-SnowyGems=GEMSRA01,GEMSRA02,GEMSRA03,GEMSRA04
+Riparius=TIB01,TIB02,TIB03,TIB04,TIB05,TIB06,TIB07,TIB08,TIB09,TIB10,TIB11,TIB12,TIB13,TIB14,TIB15,TIB16,TIB17,TIB18,TIB19,TIB20
+Cruentus=GEM01,GEM02,GEM03,GEM04,GEM05,GEM06,GEM07,GEM08,GEM09,GEM10,GEM11,GEM12
+Vinifera=TIB2_01,TIB2_02,TIB2_03,TIB2_04,TIB2_05,TIB2_06,TIB2_07,TIB2_08,TIB2_09,TIB2_10,TIB2_11,TIB2_12,TIB2_13,TIB2_14,TIB2_15,TIB2_16,TIB2_17,TIB2_18,TIB2_19,TIB2_20
+Aboreus=TIB3_01,TIB3_02,TIB3_03,TIB3_04,TIB3_05,TIB3_06,TIB3_07,TIB3_08,TIB3_09,TIB3_10,TIB3_11,TIB3_12,TIB3_13,TIB3_14,TIB3_15,TIB3_16,TIB3_17,TIB3_18,TIB3_19,TIB3_20
