+++ src/TSMapEditor/Config/Default/UI/EditorControlsPanel.ini
@@ -13,10 +13,18 @@ $CC05=chkDrawMapWideOverlay:XNACheckBox
 $CC06=btnPlaceWaypoint:EditorButton
 $CC07=btnDeletionMode:EditorButton
 $CC08=btnFrameworkMode:EditorButton
-$CC09=btnGenerateTerrain:EditorButton
-$CC10=btnTerrainGeneratorOptions:EditorButton
-$CC11=btnDrawConnectedTiles:EditorButton
-$Width=max(getRight(chkAutoLAT), getRight(chkOnlyPaintOnClearGround), getRight(chkDrawMapWideOverlay)) + EMPTY_SPACE_SIDES
+$CC09=btn2DMode:EditorButton
+$CC10=btnRaiseGround:EditorButton
+$CC11=btnLowerGround:EditorButton
+$CC12=btnRaiseGroundSteep:EditorButton
+$CC13=btnLowerGroundSteep:EditorButton
+$CC14=btnRaiseCells:EditorButton
+$CC15=btnLowerCells:EditorButton
+$CC16=btnFlattenGround:EditorButton
+$CC17=btnGenerateTerrain:EditorButton
+$CC18=btnTerrainGeneratorOptions:EditorButton
+$CC19=btnDrawConnectedTiles:EditorButton
+$Width=max(getRight(chkAutoLAT), getRight(chkOnlyPaintOnClearGround), getRight(chkDrawMapWideOverlay), getRight(btnDrawConnectedTiles)) + EMPTY_SPACE_SIDES
 $Height=getBottom(btnPlaceWaypoint) + EMPTY_SPACE_BOTTOM
 
 
@@ -32,7 +40,7 @@ $Width=110
 
 [LATPanel]
 $Width=400
-$Height=(CELL_HEIGHT * 2) + EMPTY_SPACE_TOP
+$Height=CELL_HEIGHT
 $Y=VERTICAL_SPACING / 2
 $X=getRight(ddBrushSize) + HORIZONTAL_SPACING
 DrawBorders=no
@@ -54,7 +62,7 @@ $Text=translate(Map-Wide Overlay)
 
 [btnPlaceWaypoint]
 $X=EMPTY_SPACE_SIDES
-$Y=getBottom(ddBrushSize) + EMPTY_SPACE_TOP
+$Y=getBottom(LATPanel) + EMPTY_SPACE_TOP
 $Width=BUTTON_HEIGHT
 $ToolTip=translate(Place Waypoint)
 ExtraTexture=ToolIcons/placewaypoint.png
@@ -73,8 +81,64 @@ $Width=getWidth(btnPlaceWaypoint)
 $ToolTip=translate(Toggle Framework Mode (Marble Madness))
 ExtraTexture=ToolIcons/frameworkmode.png
 
-[btnGenerateTerrain]
+[btn2DMode]
 $X=getRight(btnFrameworkMode) + HORIZONTAL_SPACING
+$Y=getY(btnPlaceWaypoint)
+$Width=getWidth(btnPlaceWaypoint)
+$ToolTip=translate(Toggle 2D Mode)
+ExtraTexture=ToolIcons/2dmode.png
+
+[btnRaiseGround]
+$X=getRight(btn2DMode) + (HORIZONTAL_SPACING * 2)
+$Y=getY(btnPlaceWaypoint)
+$Width=getWidth(btnPlaceWaypoint)
+$ToolTip=translate(Raise Ground (Non-Steep Ramps / FinalSun Mode))
+ExtraTexture=ToolIcons/raiseground.png
+
+[btnLowerGround]
+$X=getRight(btnRaiseGround) + HORIZONTAL_SPACING
+$Y=getY(btnPlaceWaypoint)
+$Width=getWidth(btnPlaceWaypoint)
+$ToolTip=translate(Lower Ground (Non-Steep Ramps / FinalSun Mode))
+ExtraTexture=ToolIcons/lowerground.png
+
+[btnRaiseGroundSteep]
+$X=getRight(btnLowerGround) + HORIZONTAL_SPACING
+$Y=getY(btnPlaceWaypoint)
+$Width=getWidth(btnPlaceWaypoint)
+$ToolTip=translate(Raise Ground (Steep Ramps))
+ExtraTexture=ToolIcons/raiseground_steep.png
+
+[btnLowerGroundSteep]
+$X=getRight(btnRaiseGroundSteep) + HORIZONTAL_SPACING
+$Y=getY(btnPlaceWaypoint)
+$Width=getWidth(btnPlaceWaypoint)
+$ToolTip=translate(Lower Ground (Steep Ramps))
+ExtraTexture=ToolIcons/lowerground_steep.png
+
+[btnRaiseCells]
+$X=getRight(btnLowerGroundSteep) + HORIZONTAL_SPACING
+$Y=getY(btnPlaceWaypoint)
+$Width=getWidth(btnPlaceWaypoint)
+$ToolTip=translate(Raise Individual Cells)
+ExtraTexture=ToolIcons/raisecells.png
+
+[btnLowerCells]
+$X=getRight(btnRaiseCells) + HORIZONTAL_SPACING
+$Y=getY(btnPlaceWaypoint)
+$Width=getWidth(btnPlaceWaypoint)
+$ToolTip=translate(Lower Individual Cells)
+ExtraTexture=ToolIcons/lowercells.png
+
+[btnFlattenGround]
+$X=getRight(btnLowerCells) + HORIZONTAL_SPACING
+$Y=getY(btnPlaceWaypoint)
+$Width=getWidth(btnPlaceWaypoint)
+$ToolTip=translate(Flatten Ground)
+ExtraTexture=ToolIcons/flattenground.png
+
+[btnGenerateTerrain]
+$X=getRight(btnFlattenGround) + (HORIZONTAL_SPACING * 2)
 $Y=getY(btnPlaceWaypoint)
 $Width=getWidth(btnPlaceWaypoint)
 $ToolTip=translate(Generate Terrain)
