+++ src/TSMapEditor/Config/Default/UI/UIConstants.ini
@@ -4,4 +4,4 @@
 ; This file defines various user-interface related constants used in the editor.
 
 [UI]
-UITreeViewLineHeight=20
+UITreeViewLineHeight=24
