+++ src/TSMapEditor/Config/Default/UI/Windows/AITriggersWindow.ini
@@ -181,6 +181,9 @@ Option2=1 House Owns
 Option3=2 Enemy On Yellow Power
 Option4=3 Enemy On Red Power
 Option5=4 Enemy Has X Credits
+Option6=5 Iron Curtain (close to) ready
+Option7=6 Chronosphere (close to) ready
+Option8=7 Neutral Owns
 
 [lblConditionType]
 $X=getX(lblSelectedAITrigger)
