+++ src/TSMapEditor/Config/Default/UI/Windows/CreateNewMapWindow.ini
@@ -24,7 +24,11 @@ $X=150
 $Width=getWidth(CreateNewMapWindow) - getX(ddTheater) - EMPTY_SPACE_SIDES
 $Y=getBottom(lblHeader) + EMPTY_SPACE_TOP
 Option0=TEMPERATE
-Option1=DESERT
+Option1=SNOW
+Option2=URBAN
+Option3=DESERT
+Option4=NEWURBAN
+Option5=LUNAR
 
 [lblTheater]
 $X=EMPTY_SPACE_SIDES
