+++ src/TSMapEditor/Config/Default/UI/Windows/LightingSettingsWindow.ini
@@ -346,8 +346,8 @@ $Y=getY(lblLightingPreview) - 1
 $Width=getWidth(tbAmbientNormal)
 Option0=No Lighting
 Option1=Normal
-Option2=Ion Storm ; Lighting Storm in YR
-;Option3=Dominator YR only
+Option2=Lightning Storm
+Option3=Dominator
 
 [btnApply]
 Width=100
