--- src/TSMapEditor/Config/MapCode/Enable Blue Tiberium Spreader Lighting.ini
@@ -1,12 +1,0 @@
-[$Editor]
-Confirmation=This script will enable blue lighting for blue tiberium spreaders. Continue?
-Success=Code for enabling blue tiberium spreader lighting has been added to the map.
-
-[TIBSPRDV]
-LightRedTint=0.0
-LightBlueTint=0.3
-LightGreenTint=0.0
-LightIntensity=0.003
-LightVisibility=1500
-
-
