--- src/TSMapEditor/Config/MapCode/Enable Green Tiberium Spreader Lighting.ini
@@ -1,26 +1,0 @@
-[$Editor]
-Confirmation=This script will enable green lighting for green tiberium spreaders. Continue?
-Success=Code for enabling green tiberium spreader lighting has been added to the map.
-
-[TIBSPRDR]
-LightRedTint=0.0
-LightBlueTint=0.0
-LightGreenTint=0.35
-LightIntensity=0.003
-LightVisibility=1500
-
-[TIBSPRDRL]
-LightRedTint=0.0
-LightBlueTint=0.0
-LightGreenTint=0.35
-LightIntensity=0.003
-LightVisibility=1500
-
-[TIBSPRDRH]
-LightRedTint=0.0
-LightBlueTint=0.0
-LightGreenTint=0.35
-LightIntensity=0.003
-LightVisibility=1500
-
-
