--- src/TSMapEditor/Config/MapCode/Enable Ore Spreader Lighting.ini
@@ -1,12 +1,0 @@
-[$Editor]
-Confirmation=This script will enable yellow lighting for ore spreaders. Continue?
-Success=Code for enabling ore spreader lighting has been added to the map.
-
-[TIBSPRDA]
-LightRedTint=0.3
-LightBlueTint=0.0
-LightGreenTint=0.3
-LightIntensity=0.003
-LightVisibility=1500
-
-
