--- src/TSMapEditor/Config/MapCode/Enable Spreading Ore.ini
@@ -1,23 +1,0 @@
-[$Editor]
-Confirmation=This script will enable spreading ore for the map. Do not use this maps that have scrap debris placed on them. Continue?
-Success=Code for enabling spreading ore has been added to the map.
-
-[Tiberiums]
-0=Riparius
-1=Cruentus
-2=Vinifera
-3=Aboreus
-
-[Aboreus]
-Name=Ore
-Image=4
-Value=25
-Growth=1000
-GrowthPercentage=10
-Spread=1000
-SpreadPercentage=2
-Power=1
-Color=DarkGold
-Debris=VINCRYS1,VINCRYS2,VINCRYS3,VINCRYS4
-
-
