--- src/TSMapEditor/Config/MapCode/Mixed Gems Only Use Blue-Green Colors.ini
@@ -1,16 +1,0 @@
-﻿[$Editor]
-Confirmation=This script will change the colors of the mixed gems to green/blue (note: the effects won't be visible in the map editor before you restart it or re-load the map). Continue?
-Success=Code for changing the image of the mixed gems has been added to the map.
-
-[GEMMIX01]
-Image=GEMGB01
-
-[GEMMIX02]
-Image=GEMGB02
-
-[GEMMIX03]
-Image=GEMGB03
-
-[GEMMIX04]
-Image=GEMGB04
-
