--- src/TSMapEditor/Config/MapCode/Replace Tiberium With Ore.ini
@@ -1,66 +1,0 @@
-[$Editor]
-Confirmation=This script will change the appearance of green tiberium to ore (note: the effects won't be visible in the map editor until you reload the map). Continue?
-Success=Code for changing the image of green tiberium has been added to the map.
-
-[Tiberiums]
-0=Riparius
-1=Cruentus
-2=Vinifera
-3=Aboreus
-
-[Riparius]
-Name=Ore
-Power=0
-
-[OREMINE]
-TiberiumToSpawn=0
-
-;[OREMINEA]
-;TiberiumToSpawn=0
-
-[OREMINE2]
-TiberiumToSpawn=0
-
-[OREMINE3]
-TiberiumToSpawn=0
-
-[TIBSPRDA]
-Image=TIBSPRDR3
-
-[RTIB01]
-Image=ZORE01
-
-[RTIB02]
-Image=ZORE02
-
-[RTIB03]
-Image=ZORE03
-
-[RTIB04]
-Image=ZORE04
-
-[RTIB05]
-Image=ZORE01
-
-[RTIB06]
-Image=ZORE02
-
-[RTIB07]
-Image=ZORE03
-
-[RTIB08]
-Image=ZORE04
-
-[RTIB09]
-Image=ZORE01
-
-[RTIB10]
-Image=ZORE02
-
-[RTIB11]
-Image=ZORE03
-
-[RTIB12]
-Image=ZORE04
-
-
