--- src/TSMapEditor/Config/MapCode/Replace Tiberium With Snowy Ore.ini
@@ -1,56 +1,0 @@
-[$Editor]
-Confirmation=This script will change the appearance of green tiberium to snowy ore (note: the effects won't be visible in the map editor until you reload the map). Continue?
-Success=Code for changing the image of green tiberium has been added to the map.
-
-[Tiberiums]
-0=Riparius
-1=Cruentus
-2=Vinifera
-3=Aboreus
-
-[Riparius]
-Name=Ore
-Power=0
-
-[OREMINEA]
-TiberiumToSpawn=0
-
-[TIBSPRDA]
-Image=TIBSPRDR3
-
-[RTIB01]
-Image=ZORESNO01
-
-[RTIB02]
-Image=ZORESNO02
-
-[RTIB03]
-Image=ZORESNO03
-
-[RTIB04]
-Image=ZORESNO04
-
-[RTIB05]
-Image=ZORESNO01
-
-[RTIB06]
-Image=ZORESNO02
-
-[RTIB07]
-Image=ZORESNO03
-
-[RTIB08]
-Image=ZORESNO04
-
-[RTIB09]
-Image=ZORESNO01
-
-[RTIB10]
-Image=ZORESNO02
-
-[RTIB11]
-Image=ZORESNO03
-
-[RTIB12]
-Image=ZORESNO04
-
