--- src/TSMapEditor/Config/MapCode/Snowy Mixed Gems.ini
@@ -1,41 +1,0 @@
-[$Editor]
-Confirmation=This script will change the appearance of the mixed gems to a snowy version (note: the effects won't be visible in the map editor until you reload the map). Continue?
-Success=Code for changing the image of the gems has been added to the map.
-
-[MIXGEM01]
-Image=GEMSRA01
-
-[MIXGEM02]
-Image=GEMSRA02
-
-[MIXGEM03]
-Image=GEMSRA03
-
-[MIXGEM04]
-Image=GEMSRA04
-
-[GRNGEM01]
-Image=GEMSRA01
-
-[GRNGEM02]
-Image=GEMSRA02
-
-[GRNGEM03]
-Image=GEMSRA03
-
-[GRNGEM04]
-Image=GEMSRA04
-
-[BLUGEM01]
-Image=GEMSRA01
-
-[BLUGEM02]
-Image=GEMSRA02
-
-[BLUGEM03]
-Image=GEMSRA03
-
-[BLUGEM04]
-Image=GEMSRA04
-
-
