--- src/TSMapEditor/Config/MapCode/Snowy Ore.ini
@@ -1,41 +1,0 @@
-﻿[$Editor]
-Confirmation=This INI file will change the appearance of ore to a snowy version (note: the effects won't be visible in the map editor before you restart it or re-load the map). Continue?
-Success=Code for changing the image of ore has been added to the map.
-
-[ZORE01]
-Image=ZORESNO01
-
-[ZORE02]
-Image=ZORESNO02
-
-[ZORE03]
-Image=ZORESNO03
-
-[ZORE04]
-Image=ZORESNO04
-
-[ZORE05]
-Image=ZORESNO01
-
-[ZORE06]
-Image=ZORESNO02
-
-[ZORE07]
-Image=ZORESNO03
-
-[ZORE08]
-Image=ZORESNO04
-
-[ZORE09]
-Image=ZORESNO01
-
-[ZORE10]
-Image=ZORESNO02
-
-[ZORE11]
-Image=ZORESNO03
-
-[ZORE12]
-Image=ZORESNO04
-
-
