--- src/TSMapEditor/Config/MapCode/Snowy Smudges.ini
@@ -1,215 +1,0 @@
-[$Editor]
-Confirmation=This script will change the appearance of smudges created by weapon impacts to a snowy version on this map. Continue?
-Success=Code for disabling normal smudges and enabling the snowy ones has been added to the map.
-
-[CR1A]
-Crater=no
-
-[CR1B]
-Crater=no
-
-[CR1C]
-Crater=no
-
-[CR1D]
-Crater=no
-
-[CR1E]
-Crater=no
-
-[CR2A]
-Crater=no
-
-[CR2B]
-Crater=no
-
-[CR2C]
-Crater=no
-
-[CR2D]
-Crater=no
-
-[CR2E]
-Crater=no
-
-[CR3A]
-Crater=no
-
-[CR3B]
-Crater=no
-
-[CR3C]
-Crater=no
-
-[CR3D]
-Crater=no
-
-[CR3E]
-Crater=no
-
-[CR4A]
-Crater=no
-
-[CR4B]
-Crater=no
-
-[CR4C]
-Crater=no
-
-[CR4D]
-Crater=no
-
-[CR4E]
-Crater=no
-
-[CR5A]
-Crater=no
-
-[CR5B]
-Crater=no
-
-[CR5C]
-Crater=no
-
-[CR5D]
-Crater=no
-
-[CR5E]
-Crater=no
-
-[CR6A]
-Crater=no
-
-[CR6B]
-Crater=no
-
-[CR6C]
-Crater=no
-
-[CR6D]
-Crater=no
-
-[CR6E]
-Crater=no
-
-[CRS1A]
-Crater=yes
-
-[CRS1B]
-Crater=yes
-
-[CRS1C]
-Crater=yes
-
-[CRS1D]
-Crater=yes
-
-[CRS1E]
-Crater=yes
-
-[CRS2A]
-Crater=yes
-
-[CRS2B]
-Crater=yes
-
-[CRS2C]
-Crater=yes
-
-[CRS2D]
-Crater=yes
-
-[CRS2E]
-Crater=yes
-
-[CRS3A]
-Crater=yes
-
-[CRS3B]
-Crater=yes
-
-[CRS3C]
-Crater=yes
-
-[CRS3D]
-Crater=yes
-
-[CRS3E]
-Crater=yes
-
-[CRS4A]
-Crater=yes
-
-[CRS4B]
-Crater=yes
-
-[CRS4C]
-Crater=yes
-
-[CRS4D]
-Crater=yes
-
-[CRS4E]
-Crater=yes
-
-[CRS5A]
-Crater=yes
-
-[CRS5B]
-Crater=yes
-
-[CRS5C]
-Crater=yes
-
-[CRS5D]
-Crater=yes
-
-[CRS5E]
-Crater=yes
-
-[CRS6A]
-Crater=yes
-
-[CRS6B]
-Crater=yes
-
-[CRS6C]
-Crater=yes
-
-[CRS6D]
-Crater=yes
-
-[CRS6E]
-Crater=yes
-
-[SC1]
-Burn=no
-
-[SC2]
-Burn=no
-
-[SC4]
-Burn=no
-
-[SC5]
-Burn=no
-
-[SC6]
-Burn=no
-
-[SCS1]
-Burn=yes
-
-[SCS2]
-Burn=yes
-
-[SCS4]
-Burn=yes
-
-[SCS5]
-Burn=yes
-
-[SCS6]
-Burn=yes
-
-
