--- src/TSMapEditor/Config/MapCode/Snowy Tiberium.ini
@@ -1,77 +1,0 @@
-﻿[$Editor]
-Confirmation=This INI file will change the appearance of tiberium to a snowy version (note: the effects won't be visible in the map editor before you restart it or re-load the map). Continue?
-Success=Code for changing the image of tiberium has been added to the map.
-
-[RTIB01]
-Image=TIB1S_01
-
-[RTIB02]
-Image=TIB1S_02
-
-[RTIB03]
-Image=TIB1S_03
-
-[RTIB04]
-Image=TIB1S_04
-
-[RTIB05]
-Image=TIB1S_05
-
-[RTIB06]
-Image=TIB1S_06
-
-[RTIB07]
-Image=TIB1S_07
-
-[RTIB08]
-Image=TIB1S_08
-
-[RTIB09]
-Image=TIB1S_09
-
-[RTIB10]
-Image=TIB1S_10
-
-[RTIB11]
-Image=TIB1S_11
-
-[RTIB12]
-Image=TIB1S_12
-
-[QTIB01]
-Image=TIB2S_01
-
-[QTIB02]
-Image=TIB2S_02
-
-[QTIB03]
-Image=TIB2S_03
-
-[QTIB04]
-Image=TIB2S_04
-
-[QTIB05]
-Image=TIB2S_05
-
-[QTIB06]
-Image=TIB2S_06
-
-[QTIB07]
-Image=TIB2S_07
-
-[QTIB08]
-Image=TIB2S_08
-
-[QTIB09]
-Image=TIB2S_09
-
-[QTIB10]
-Image=TIB2S_10
-
-[QTIB11]
-Image=TIB2S_11
-
-[QTIB12]
-Image=TIB2S_12
-
-
