+++ src/TSMapEditor/TSMapEditor.csproj
@@ -27,14 +27,10 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Remove="Config\Scripts\Activate Assign Mission Cursor Action.cs" />
-    <Compile Remove="Config\Scripts\Apply Animated Water.cs" />
     <Compile Remove="Config\Scripts\Disable All Debug Triggers.cs" />
     <Compile Remove="Config\Scripts\Make Civilian Vehicles Sleep.cs" />
     <Compile Remove="Config\Scripts\Remove All Terrain Objects.cs" />
-    <Compile Remove="Config\Scripts\Replace Autumn Grass With Tall Grass.cs" />
-    <Compile Remove="Config\Scripts\Replace Autumn Jungle Grass With Summer Grass.cs" />
-    <Compile Remove="Config\Scripts\Replace Autumn Trees With Summer Trees.cs" />
-    <Compile Remove="Config\Scripts\Replace Frost With Clear.cs" />
+    <Compile Remove="Config\Scripts\Smoothen Water.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Update="Config\DefaultSettings.ini">
@@ -487,6 +483,12 @@
     <None Update="Content\trackbarButton.png">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
+    <None Update="marble.mix">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+	<None Update="WAECache.mix">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="ImeSharp" Version="1.4.1" />
@@ -505,7 +507,7 @@
     <None Include="Config\Scripts\Activate Assign Mission Cursor Action.cs">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Include="Config\Scripts\Apply Animated Water.cs">
+    <None Include="Config\Scripts\Smoothen Water.cs">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
     <None Include="Config\Scripts\Count Credits On Map.cs">
@@ -516,18 +518,6 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
     <None Include="Config\Scripts\Remove All Terrain Objects.cs">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="Config\Scripts\Replace Autumn Trees With Summer Trees.cs">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="Config\Scripts\Replace Autumn Grass With Tall Grass.cs">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="Config\Scripts\Replace Autumn Jungle Grass With Summer Grass.cs">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="Config\Scripts\Replace Frost With Clear.cs">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
